pyproject.toml
[project]
name = "agnia-smart-digest"
version = "0.1.0"
description = "Smart Digest Actions and Dynamic Planner for Agnia Assistant at Innopolis AES"
authors = [
    { name = "Anton Kudryavtsev", email = "antonkudryavtsevdoem@gmail.com" },
]
dependencies = [
    "fastapi>=0.111.0",
    "pydantic>=2.4.1",
    "aiohttp>=3.9.5",
    "httpx>=0.27.0",
    "loguru>=0.7.2",
    "uvicorn>=0.30.1",
    "pydantic-settings>=2.3.4",
    "requests>=2.32.3",
    "websockets>=12.0",
    "aiogram>=3.8.0",
    "python-weather>=2.0.3",
    "beautifulsoup4>=4.12.3",
    "nltk>=3.8.1",
    "newsapi-python>=0.2.7",
    "selenium>=4.22.0",
    "psutil>=6.0.0",
    "numpy>=1",
    "lexrank>=0.1.0",
    "humanfriendly>=10.0",
    "chromadb>=0.5.3",
]
readme = "README.md"
requires-python = ">= 3.11"

[project.scripts]
"agnia-smart-digest" = "agnia_smart_digest:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.2",
    "pytest-aio>=1.8.1",
    "pytest-sugar>=1.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.4.2",
    "ruff>=0.5.1",
    "mypy>=1.10.1",
    "isort>=5.13.2",
    "requests>=2.32.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/agnia_smart_digest"]
name = "agnia-smart-digest"
version = "0.1.0"
